arch/arm64: target cortex-a75 when building with LLVM/Clang
The LLVM backend has had support for AArch64 Cortex-A75 and
Cortex-A55 since 2017 [1], and from AArch64TargetParser.def we
can see that cortex-a75 and cortex-a55 have the same ID, FPU, and
default extensions. Therefore, setting mcpu to cortex-a75 or a55
should have the same effect. The supported targets in LLVM
can also be queried by running:

$ llc -march=arm64 -mattr=help

Regarding mcpu, march, and mtune, ARM recommends setting mcpu
only while avoiding march and mtune [2], as long as the binary
is intended to be optimized for a specific architecture, which is
exactly the case of this kernel. Qualcomm SDM845 also has the
crypto and crc32 extensions, so add those as well to avoid
issues on optimized crypto assembly routines.

[1] https://github.com/llvm/llvm-project/commit/b252ffd2cca9e9f3ed65b8912603abc1f006ac5a
[2] https://community.arm.com/developer/tools-software/tools/b/tools-software-ides-blog/posts/compiler-flags-across-architectures-march-mtune-and-mcpu

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>